+++ src/main/java/video/api/java/sdk/domain/exception/ClientException.java
@@ -1,12 +1,10 @@
 package video.api.java.sdk.domain.exception;
 
-import kong.unirest.HttpResponse;
 import kong.unirest.JsonNode;
 
 public class ClientException extends ResponseException {
 
-    public ClientException(HttpResponse<JsonNode> response, String message) {
-        super(response, "Client ERROR -> Client Exception\n" + "\n Response Code : " + response.getStatus() + "\n" + message);
+    public ClientException(String message, JsonNode responseBody, int status) {
+        super(message, responseBody, status);
     }
-
 }