+++ src/main/java/video/api/java/sdk/domain/exception/ServerException.java
@@ -4,13 +4,7 @@ import kong.unirest.HttpResponse;
 import kong.unirest.JsonNode;
 
 public class ServerException extends ResponseException {
-
-
-    public ServerException(HttpResponse<JsonNode> response, String message) {
-        super(response, "Server ERROR ->" + "\n Response Code : " + response.getStatus() + "\n" + message);
-
-
+    public ServerException(String message, JsonNode responseBody, int status) {
+        super(message, responseBody, status);
     }
-
-
 }