+++ src/main/java/video/api/java/sdk/domain/player/PlayerClient.java
@@ -3,6 +3,8 @@ package video.api.java.sdk.domain.player;
 import video.api.java.sdk.domain.QueryParams;
 import video.api.java.sdk.domain.exception.ResponseException;
 
+import java.io.File;
+
 
 public interface PlayerClient {
 
@@ -12,7 +14,7 @@ public interface PlayerClient {
 
     Player update(Player player) throws ResponseException;
 
-    Player uploadLogo(String playerId, String logoSource, String link) throws ResponseException;
+    Player uploadLogo(String playerId, File file, String link) throws ResponseException;
 
     void delete(String playerId) throws ResponseException;